Add schedule groups to SecureRooms
