+++ app/views/product_access_groups/_headers.html.haml
@@ -1,7 +1,7 @@
 = content_for :h1 do
-  =h @facility
+  = current_facility
 = content_for :sidebar do
-  = render :partial => 'admin/shared/sidenav_product', :locals => { :sidenav_tab => 'instruments' }
+  = render "admin/shared/sidenav_product", sidenav_tab: @product.product_type
 = content_for :tabnav do
-  = render :partial => 'admin/shared/tabnav_product', :locals => {:secondary_tab => 'restriction_levels'}
-%h2= @instrument
+  = render "admin/shared/tabnav_product", secondary_tab: "restriction_levels"
+%h2= @product
