+++ app/views/product_access_groups/new.html.haml
@@ -1,5 +1,5 @@
 = render :partial => "headers"
 %h3 New #{ProductAccessGroup.model_name.human}
-= simple_form_for([current_facility, @instrument, @product_access_group]) do |f|
+= simple_form_for([current_facility, @product, @product_access_group]) do |f|
   = f.error_messages
   = render :partial => "fields", :locals => {:f => f}
