+++ app/views/product_users/index.html.haml
@@ -23,7 +23,7 @@
 - elsif @product_users.empty?
   %p.notice= t(".no_users_approved")
 - else
-  = form_for @product, url: facility_instrument_update_restrictions_path(current_facility, @product), method: :put do |f|
+  = form_for @product, url: [current_facility, @product, :update_restrictions], method: :put do |f|
     %table.table.table-striped.table-hover
       %thead
         %tr
